Airplane service
Created separation of concerns.I have created airplane service that will
take care of all the airplane's business logic. Controller just pass on
the request to the service.I separated the routes in v1, v2 etc so that
we can maintain the code more efficiently. Airplane model is created. I
also generated a Migration file for the Database so that we can track
the schema changes of the database.